cell quench data

A table containing experimental data for informing how cells quench fluorescence of E. coli cells.

cell_quench_data

Format

A data frame with 34 rows and 12 variables:

channel_name

name of filter set used

channel_ex

excitation filter

channel_em

emission filter

media

media

calibrant

calibrant

protein

protein

cells_od600

accurate OD600 cm-1 values

cells_od700

accurate OD700 cm-1 values

volume

volume in ul

measure

filter set plus gain

gain

gain

autof_norm_value

fluorescence values normalised to remove the contribution of cellular autofluorescence

norm_fc_value

fold change between the fluorescence sample and its relevant negative control without cells