Normalise fluorescence against negative well autofluorescence

Used by process_plate function for fluorescence normalisation. Remains virtually unchanged from flopr::flu_norm, except that af_model can be set to NULL to normalise to blank wells instead, and files are saved to a specified outfolder.

flu_norm(
  pr_data,
  neg_well,
  blank_well,
  flu_name,
  af_model,
  data_csv,
  outfolder
)

Arguments

pr_data
a long data.frame containing you plate reader data with OD normalised

neg_well
the well coordinates of a non-fluorescent control

blank_well
the well coordinates of a media blank

flu_name
the column name of the fluorescence chanel to normalise

af_model
model used to fit negative control autofluorescence. For now these include "polynomial", "invers_poly", "exponential", "spline" or "loess". If set to NULL, no model is used, and fluorescence is normalised akin to OD: by subtracting the value for the blanks.

data_csv
path to the original data. Used for saving normalisation curve plots.

outfolder
path to folder where output files should be saved. Defaults to current working directory.

Value