Get FP concentrations using BCA method

Get protein's concentration from a dilution series measured with the BCA assay. Takes two input data sets, the BCA assay data and the A562 baseline data. The A562 baseline data is necessary for proteins that might naturally absorb in this range. Script normalises BCA assay data for A562 baseline, then to blank values of BCA. Data from the BSA standards (identified by protein column containing the word "BSA"), are used to construct a standard curve of concentration in ng/ul vs normalised A562 values, which is used to predict FP concentrations at each dilution. The FP's concentration vs dilution values are then used to predict the FP concentration at each dilution in one of two ways. Where option is set to fit, a linear model is fitted between FP concentration and dilution, and the fitted values are exported. Where option is set to highest, FP concentrations are taken only from the highest concentration/dilution specified and exported.

get_conc_bca(
  microbca_data_csv,
  a562_baseline_csv = NULL,
  calibr,
  buffer,
  protein_seq,
  option = "highest",
  outfolder = "."
)

Arguments

microbca_data_csv
path of the csv file of your microBCA data. required.

a562_baseline_csv
path of the csv file of your A562 baseline data. Optional. If data is missing, use NULL. If NULL is specified, the value of 0 is assigned as baseline for all wells. Default is NULL.

calibr
string specifying the value of the 'calibrant' column to assess with this function. Function subsets the data by the value specified here. This works by taking all the rows with specified string in the calibrant column and discarding all other rows (which means that the blanks relevant to the specified calibrant need to be specified as calibrant = calibr, protein = "none", otherwise they will be removed).

buffer
string specifying the value of the 'media' column to assess with this function. Function subsets the data by the value specified here.

protein_seq
character string of protein sequence using 1-letter code. Required for MW calculation.

option
string specifying how to choose the predicted concentration to use. Default is "highest", in which the mean predicted conc of the highest dilution (ie. neat) is used, and is multiplied by the dilution to determine the concentration of the other dilutions. The alternative, "fit", fits a y=mx linear model and uses that for the concentration determination.

outfolder
path to the folder where output files should be saved. Defaults to the current working directory.

Examples

bca_concs <- get_conc_bca(microbca_data_csv = "bca_data_parsed.csv", a562_baseline_csv = "a562_data_parsed.csv", calibr = "mCherry", buffer = "T5N15_pi", protein_seq = protein_seq, option = "highest", outfolder = "protquant_microbca/mCherry_T5N15pi")