Get a protein's molar extinction coefficient at A280

The purpose of this function is to work out the theoretical molar extinction coefficient of a protein at 280nm (EC280, M-1cm-1) using only the protein's primary sequence, using the ProtParam method (Pace values). A full explanation of the method can be found at https://web.expasy.org/protparam/protparam-doc.html

get_extcoeff_A280(
  protein,
  disulphides = FALSE,
  showWarnings = TRUE,
  showMessages = TRUE,
  protein_name = "-",
  buffer = "-",
  mol_weight = NULL,
  save = TRUE,
  outfolder = "."
)

Arguments

protein
character string of protein sequence using 1-letter code

disulphides
logical. Does protein have disulphides?

showWarnings
logical. Should function show warnings?

showMessages
logical. Should function show messages?

protein_name
character string of protein name. Optional.

buffer
character string of buffer. Optional.

mol_weight
numerical value for molecular weight (g/mol). Optional. If specified the function gives extinction coefficients for 1% (10mg/ml) and 0.1% (1mg/ml) solutions too.

save
logical. Should function save csv file of output?

outfolder
path to folder where output files should be saved. Defaults to current working directory.