View k-factors

Displays the data tables used for k-factor calculations in get_kfactors(), along with notes.

Usage

view_kfactors()

Examples

view_kfactors()
#> 
#> Data from https://static.thermoscientific.com/images/D20827~.pdf :
#>            buffer concentration units         description kfactor fold_change
#> 1           water          0.00  none               Water   0.173   1.0000000
#> 2    Na_phosphate          0.10     M  Na_phosphate_100mM   0.170   0.9826590
#> 3            NaOH          0.02     M           NaOH_20mM   0.169   0.9768786
#> 4  sulphuric_acid          0.02     M sulphuric_acid_20mM   0.169   0.9768786
#> 5       TE_buffer          0.01     M      TE_buffer_10mM   0.169   0.9768786
#> 6            NaCl          0.90    pc           NaCl_0.9%   0.168   0.9710983
#> 7             TBS          0.05     M            TBS_50mM   0.166   0.9595376
#> 8    Na_phosphate          0.08     M   Na_phosphate_80mM   0.166   0.9595376
#> 9        Tris-HCl          1.00     M         Tris-HCl_1M   0.157   0.9075145
#> 10          Tween          0.10    pc          Tween_0.1%   0.155   0.8959538
#> 11 sulphuric_acid          2.00     M   sulphuric_acid_2M   0.154   0.8901734
#> 12           DMSO          1.00    pc             DMSO_1%   0.169   0.9768786
#> 13           DMSO          5.00    pc             DMSO_5%   0.159   0.9190751
#> 14           DMSO         10.00    pc            DMSO_10%   0.148   0.8554913
#> 15           EtOH          5.00    pc             EtOH_5%   0.159   0.9190751
#> 16           EtOH         20.00    pc            EtOH_20%   0.129   0.7456647
#> 17           EtOH         40.00    pc            EtOH_40%   0.092   0.5317919
#> 18           MeOH          5.00    pc             MeOH_5%   0.161   0.9306358
#> 19           MeOH         20.00    pc            MeOH_20%   0.133   0.7687861
#> 20           MeOH         40.00    pc            MeOH_40%   0.100   0.5780347
#> 
#>   temperature kfactor fold_change
#> 1          25   0.172       1.000
#> 2          28   0.174       1.012
#> 3          31   0.177       1.029
#> 4          34   0.179       1.041
#> 5          37   0.183       1.064
#> 6          41   0.188       1.093
#> 7          45   0.191       1.110
#> 
#> The k-factor of water (A975-A900 at 1cm) at ~25oC is 0.172.
#> Increasing the temperature has a small effect: 30oC = 3% increase to 0.176.
#> The volume of water is minimally effected by temperature, so path lengths calculated at 25oC should apply to other temperatures.
#> 
#> Most aqueous buffers with low salt will be ~ 0.170 and can therefore be approximated to the k-factor of pure water.
#> eg.
#> 153mM (0.9%) NaCl = 0.168, TE = 0.169, 50mM TBS = 0.166
#> 1M Tris-HCl = 0.157, 2M H2SO4 = 0.154
#> 
#> Other solvents, like DMSO, EtOH and MeOH have a bigger effect.
#> eg.
#> 10% DMSO = 0.148
#> 20% EtOH = 0.092, 20% MeOH = 0.100
#> 100% EtOH and MeOH have negative k-factors, and k-factors with these solvents are no longer reliable methods for calculating path lengths.