Get FP concentrations using A280 method

Get protein's concentration from a dilution series measured with an absorbance spectrum. Expects 'processed' data such as that produced by process_absorbance_spectrum(), with a file name ending _processed.csv, which contains values corrected for path length and normalised to blanks as a column called normalised_cm1_value, but retains replicate data containing positional (well) information required for exporting predicted concentrations at the end of this function. Uses get_extcoeff_a280() to get EC in M-1cm-1 and wavelength, and converts it to an EC mass extinction coefficient in (mgml)-1cm-1 using the MW (worked out from protein_seq and fpcountr::get_mw). Then the function uses the EC_A280_mgml to work out the concentration of protein in each well, using three correction methods. Instead of using the normalised data directly, the values used are based on a LOESS fit through the absorption spectra to minimise fluctuations due to noise. Finally, linear models are fitted to each concentration prediction method, and a dataframe is built, returned and saved, containing predicted concentrations according to the user's chosen correction method. Plots showing each of the analytical steps are saved concurrently.

Usage

get_conc_a280(
  protein_slug,
  protein_seq,
  buffer = "",
  processed_spectrum_csv,
  wells_to_remove = NULL,
  disulphides = FALSE,
  showWarnings = TRUE,
  showMessages = FALSE,
  corr_method = "none",
  wav_to_use1 = 340,
  wav_to_use2 = 333,
  outfolder = ""
)

Arguments

protein_slug

character string of protein name in 'slug' form to match slug of FPbase entry.

protein_seq

character string of protein sequence using 1-letter code. Required for MW calculation.

buffer

character string of buffer. Optional. Defaults to "".

processed_spectrum_csv

Path to CSV file of a processed absorbance spectrum. Processing should be done with process_absorbance_spectrum(), which corrects for path lengths and normalises to blank wells.

wells_to_remove

list of wells to remove before analysis. Defaults to NULL.

disulphides

required for calculation of A280 extinction coefficient. logical. Does protein have disulphides? Defaults to FALSE.

showWarnings

required for calculation of A280 extinction coefficient. logical. Should function show warnings? Defaults to TRUE.

showMessages

required for calculation of A280 extinction coefficient. logical. Should function show messages? Defaults to TRUE.

corr_method

string corresponding to type of correction method to use for the data to remove contribution of light scatter. Options are none, baseline and scatter. Method none applies no correction. Method baseline subtracts the absorbance value of the wavelength supplied in wav_to_use1. Method scatter subtracts a fraction of the absorbance value of the wavelength supplied in wav_to_use2 according to scatter theory (details in section).

wav_to_use1

numerical value of wavelength (nm) to use for baseline correction. Defaults to 340nm.

wav_to_use2

numerical value of wavelength (nm) to use for scatter correction. Defaults to 333nm.

outfolder

path to folder where output files should be saved.

Value

data frame of concentration data

Examples

if (FALSE) {
  a280_concs <- get_conc_a280(
    protein_slug = "mcherry", protein_seq = protein_seq, buffer = "T5N15_pi",
    processed_spectrum_csv = "abs_parsed_processed.csv",
    corr_method = "scatter", wav_to_use1 = 340, wav_to_use2 = 315,
    outfolder = "protquant_a280/mCherry_T5N15pi"
  )
}