Get k-factor of a buffer

Calculates k-factor of a given buffer/temperature combination according to data from Thermo Scientific for different buffers and temperatures. The k-factor is defined as the A975-A900 for a 1cm pathlength. The calculations are made under the assumption that the fold changes for buffers from water, and for temperatures from 25oC may be combined to get an approximate estimate of the k-factor for a given condition.

Usage

get_kfactor(
  buffer_used = "water",
  concentration_used = 0,
  temperature_used = 25
)

Arguments

buffer_used

string corresponding to buffer. Must match a buffer entry in dataset kfactors_buffers_data. Default is "water".

concentration_used

numeric value of concentration of buffer used. Default is 0 (pure water), so it needs changing if buffer isn't water. Take care to ensure the units are correct (buffers specified in M require concentrations in M not mM). This value is ignored if buffer = "water".

temperature_used

numeric value of temperature in oC. Default is 25.

Examples

get_kfactor(buffer_used = "TBS", concentration_used = 0.005, temperature_used = 30)
#> 
#> Calculating k-factor for TBS at concentration 0.005 at temperature 30 oC.
#> 
#> Reference k-factor 0.172.
#> 
#> K-factors available for given buffer: 
#>   buffer concentration units description kfactor fold_change
#> 1    TBS          0.05     M    TBS_50mM   0.166   0.9595376
#> 
#> Values used for model (kfactor ~ concentration): 
#>   buffer concentration units description kfactor fold_change
#> 1  water          0.00  none       Water   0.173   1.0000000
#> 2    TBS          0.05     M    TBS_50mM   0.166   0.9595376
#> 
#> Change in k-factor required for given buffer: 0.996.
#> 
#> Values used for model (fold_change ~ temperature): 
#>   temperature kfactor fold_change
#> 1          25   0.172       1.000
#> 2          28   0.174       1.012
#> 3          31   0.177       1.029
#> 4          34   0.179       1.041
#> 5          37   0.183       1.064
#> 6          41   0.188       1.093
#> 7          45   0.191       1.110
#> 
#> Change in k-factor required for given temperature: 1.024.
#> 
#> Overall k-factor: 0.175.
#> [1] 0.1754875